feat(vra): add cli reporter
