+++ tools/visual-regression-assert/package.json
@@ -19,7 +19,8 @@
     "pixelmatch": "^7.1.0",
     "ejs": "^3.1.7",
     "glob": "^11.0.1",
-    "find-up": "5.0.0"
+    "find-up": "5.0.0",
+    "cli-table3": "^0.6.5"
   },
   "devDependencies": {
     "@types/ejs": "3.1.5"
